% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership_tables.R
\name{membership_table}
\alias{membership_table}
\title{Reshape membership table from wide to long format}
\usage{
membership_table(
  xref,
  original_names,
  target_names,
  membership_values,
  target = NULL
)
}
\arguments{
\item{xref}{Input cross-referencing table (wide format)}

\item{original_names}{Names or codes of the original ecosystem concepts}

\item{target_names}{Names or codes of the target ecosystem concepts}

\item{membership_values}{Membership values (between 0 and 1)}

\item{target}{Filter input to focus on specific targets}
}
\value{
A membership table in long format
}
\description{
Reshape membership table from wide to long format
}
